Update on "gtest-ify JIT tests, through the letter c"
See https://github.com/pytorch/pytorch/pull/45018 for context.

Differential Revision: [D23802296](https://our.internmc.facebook.com/intern/diff/D23802296)

[ghstack-poisoned]
